Internal: introduce the `starknet-os-types` crate
Problem: we need a single crate that defines useful types for all parts
of the project. Useful types should be available from all the other
crates.

Solution: introduce the `starknet-os-types` crate. Moved the `Hash` type
there as a demonstration and preparation for upcoming PRs.
