+++ crates/starknet-os/src/storage/storage.rs
@@ -1,17 +1,12 @@
 use std::marker::PhantomData;
-use std::ops::{Deref, DerefMut};
+use std::ops::DerefMut;
 use std::sync::Arc;
 
 use blockifier::state::errors::StateError;
 use cairo_vm::Felt252;
-use num_bigint::BigUint;
-use serde::{Deserialize, Serialize};
-use starknet_api::core::{ClassHash, CompiledClassHash};
-use starknet_api::hash::StarkFelt;
-use starknet_api::StarknetApiError;
+use starknet_os_types::hash::Hash;
 use tokio::sync::Mutex;
 
-use crate::starkware_utils::commitment_tree::patricia_tree::patricia_tree::EMPTY_NODE_HASH;
 use crate::starkware_utils::serializable::{DeserializeError, Serializable, SerializeError};
 
 pub const HASH_BYTES: usize = 32;
@@ -39,101 +34,6 @@ pub trait Storage: Sync + Send {
     async fn set_value(&mut self, key: Vec<u8>, value: Vec<u8>) -> Result<(), StorageError>;
 
     fn get_value(&self, key: &[u8]) -> impl futures::Future<Output = Result<Option<Vec<u8>>, StorageError>> + Send;
-}
-
-/// Starknet hash type.
-/// Encapsulates the result of hash functions and provides conversion functions to Cairo VM
-/// and Starknet API types for convenience.
-#[derive(Debug, Clone, Hash, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(transparent)]
-pub struct Hash([u8; 32]);
-
-impl Hash {
-    pub fn empty() -> Self {
-        Self::from_bytes_be(EMPTY_NODE_HASH)
-    }
-
-    pub fn from_bytes_be(bytes: [u8; 32]) -> Self {
-        Self(bytes)
-    }
-
-    /// Builds a `Hash` from a bytes slice.
-    /// The slice length must be <= 32.
-    pub fn from_bytes_be_slice(bytes: &[u8]) -> Self {
-        let mut array = [0u8; 32];
-        let start = 32 - bytes.len();
-
-        for (i, &byte) in bytes.iter().enumerate() {
-            array[start + i] = byte;
-        }
-
-        Hash(array)
-    }
-}
-
-impl PartialEq<[u8; 32]> for Hash {
-    fn eq(&self, other: &[u8; 32]) -> bool {
-        &self.0 == other
-    }
-}
-
-impl Deref for Hash {
-    type Target = [u8];
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl From<Hash> for cairo_vm::Felt252 {
-    fn from(hash: Hash) -> Self {
-        cairo_vm::Felt252::from_bytes_be(&hash.0)
-    }
-}
-
-impl From<&Hash> for BigUint {
-    fn from(hash: &Hash) -> Self {
-        BigUint::from_bytes_be(&hash.0)
-    }
-}
-
-impl From<&BigUint> for Hash {
-    fn from(value: &BigUint) -> Self {
-        // `BigUint.to_bytes_be()` only returns the minimum amount of bytes, so we need to use
-        // `from_bytes_be_slice` for this conversion.
-        Self::from_bytes_be_slice(&value.to_bytes_be())
-    }
-}
-
-impl From<cairo_vm::Felt252> for Hash {
-    fn from(value: cairo_vm::Felt252) -> Self {
-        // This conversion is safe, BigUint is 32 bytes so this will always work.
-        Self::from_bytes_be(value.to_bytes_be())
-    }
-}
-
-impl TryFrom<Hash> for StarkFelt {
-    type Error = StarknetApiError;
-
-    fn try_from(hash: Hash) -> Result<Self, Self::Error> {
-        Self::new(hash.0)
-    }
-}
-
-impl TryFrom<Hash> for CompiledClassHash {
-    type Error = StarknetApiError;
-
-    fn try_from(hash: Hash) -> Result<Self, Self::Error> {
-        Ok(Self(hash.try_into()?))
-    }
-}
-
-impl TryFrom<Hash> for ClassHash {
-    type Error = StarknetApiError;
-
-    fn try_from(hash: Hash) -> Result<Self, Self::Error> {
-        Ok(Self(hash.try_into()?))
-    }
 }
 
 pub trait HashFunctionType {
